Reframe LQI workflow as design decision optimization
